chore: initial refactor solvers
